Make it runnable from command line

+++ src/main/scala/models/Order.scala
@@ -4,7 +4,7 @@ import java.util.Date
 
 case class Order(id: Int,
                  orderType: String,
-                 quantity: Int,
+                 quantity: Double,
                  price: Float,
                  stockName: String,
                  status: String,
